# api_spec.yaml
openapi: 3.0.3
info:
  title: CAD Agent API
  version: 1.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/jobs/{job_id}:
    get:
      summary: Get job status/result
      parameters:
        - name: job_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Job state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/JobStatus"}
...
  /v1/sessions/{session_id}/logs:
    get:
      summary: Get structured logs for a session
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - name: since
          in: query
          required: false
          schema: {type: string, format: date-time}
      responses:
        "200":
          description: Log events
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LogEventList"}
...
  /v1/artifacts/{artifact_id}:
    get:
      summary: Get a presigned download URL
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Artifact download info
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ArtifactDownload"}
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    JobStatus:
      type: object
      required: [job_id, status, session_id, user_message_id]
      properties:
        job_id: {type: string, format: uuid}
        status: {type: string, enum: [queued, started, finished, failed]}
        session_id: {type: string, format: uuid}
        user_message_id: {type: string, format: uuid}
        result:
          type: object
          nullable: true
          description: Present when finished
        error:
          type: object
          nullable: true
          description: Present when failed
    ArtifactDownload:
      type: object
      required: [artifact_id, kind, object_key, download_url, expires_at]
      properties:
        artifact_id: {type: string, format: uuid}
        kind: {type: string}
        object_key: {type: string}
        download_url: {type: string}
        expires_at: {type: string, format: date-time}
    LogEventList:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: {$ref: "#/components/schemas/LogEvent"}
    LogEvent:
      type: object
      required: [event_id, session_id, ts, type]
      properties:
        event_id: {type: string, format: uuid}
        session_id: {type: string, format: uuid}
        ts: {type: string, format: date-time}
        type: {type: string}
        payload: {type: object, additionalProperties: true}
...